requirements.txt
asgiref==3.2.3
Django==3.0
django-cors-headers==3.2.0
django-countries==5.5
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.1.1
Pillow==6.2.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
